# src/main/java/com/bootcamp/ms_credits/infrastructure/adapters/inbound/rest/dto/ChargeRequest.java
package com.bootcamp.ms_credits.infrastructure.adapters.inbound.rest.dto;

import jakarta.validation.constraints.Positive;
import lombok.Data;

@Data
public class ChargeRequest {

    @Positive
    private Double amount;

    private String description;
}